test/reactor: move non-handle tests to libflux/test
